refactor: rename variables and simplify number of files, delete dead code
