+++ app-swift/Fushigi/Models/Tags.swift
@@ -5,8 +5,11 @@
 //  Created by Tahoe Schrader on 2025/08/09.
 //
 
-// MARK: Level
+import SwiftUI
 
+// MARK: - Enums
+
+/// Politeness level filter options
 enum Level: String, CaseIterable, Identifiable {
     case all = "All Levels"
     case casual = "Casual"
@@ -30,8 +33,7 @@ enum Level: String, CaseIterable, Identifiable {
     }
 }
 
-// MARK: Context
-
+/// Usage context filter options
 enum Context: String, CaseIterable, Identifiable {
     case all = "All Contexts"
     case spoken = "Spoken"
@@ -51,8 +53,7 @@ enum Context: String, CaseIterable, Identifiable {
     }
 }
 
-// MARK: Fun Mode
-
+/// Language variant filter options
 enum FunMode: String, CaseIterable, Identifiable {
     case none = "No Extras"
     case slang = "Slang"
@@ -70,8 +71,7 @@ enum FunMode: String, CaseIterable, Identifiable {
     }
 }
 
-// MARK: Source Mode
-
+/// Grammar sourcing algorithm options
 enum SourceMode: String, CaseIterable, Identifiable {
     case random = "Random"
     case srs = "SRS"
@@ -93,4 +93,26 @@ enum SourceMode: String, CaseIterable, Identifiable {
         case .srs: "brain.head.profile"
         }
     }
+}
+
+// MARK: - Helper Functions
+
+/// Create colored text from array of tags
+func coloredTagsText(tags: [String]) -> Text {
+    var combinedText = Text("")
+
+    for (index, tag) in tags.enumerated() {
+        let coloredText = Text(tag)
+            .font(.caption)
+            .foregroundColor(index.isMultiple(of: 2) ? .primary : .secondary)
+
+        // swiftlint:disable:next shorthand_operator
+        combinedText = combinedText + coloredText
+
+        if index < tags.count - 1 {
+            // swiftlint:disable:next shorthand_operator
+            combinedText = combinedText + Text("   ").font(.caption).foregroundColor(.primary)
+        }
+    }
+    return combinedText
 }
