+++ app-swift/Fushigi/Subcomponents/DetailedGrammar.swift
@@ -1,5 +1,5 @@
 //
-//  InspectorView.swift
+//  DetailedGrammar.swift
 //  fushigi
 //
 //  Created by Tahoe Schrader on 2025/08/08.
@@ -7,11 +7,23 @@
 
 import SwiftUI
 
-struct InspectorView: View {
+// MARK: - DetailedGrammar
+
+/// Detailed view displaying comprehensive grammar point information
+struct DetailedGrammar: View {
+    /// Grammar point to display
     let grammarPoint: GrammarPointModel
+
+    /// Controls view presentation state
     @Binding var isPresented: Bool
+
+    /// Currently selected grammar point ID
     @Binding var selectedGrammarID: UUID?
+
+    /// Layout mode indicator for responsive design
     let isCompact: Bool
+
+    // MARK: - Main View
 
     var body: some View {
         VStack(alignment: .leading, spacing: 16) {
@@ -40,13 +52,15 @@ struct InspectorView: View {
     }
 }
 
+// MARK: - Previews
+
 #Preview {
     @Previewable @Environment(\.horizontalSizeClass) var horizontalSizeClass
     var isCompact: Bool {
         horizontalSizeClass == .compact
     }
     PreviewHelper.withStore { store in
-        InspectorView(
+        DetailedGrammar(
             grammarPoint: store.grammarItems.last!,
             isPresented: .constant(true),
             selectedGrammarID: .constant(store.grammarItems.first!.id),
