+++ app-swift/Fushigi/Utilities/Extensions.swift
@@ -5,11 +5,12 @@
 //  Created by Tahoe Schrader on 2025/08/20.
 //
 
-import Foundation
 import SwiftUI
 
+// MARK: - Extensions
+
 extension JSONDecoder {
-    /// A JSONDecoder configured to decode ISO8601 dates with fractional seconds from timestamptz
+    /// JSONDecoder configured for ISO8601 dates with fractional seconds from timestamptz
     static var iso8601withFractionalSeconds: JSONDecoder {
         let decoder = JSONDecoder()
         let formatter = ISO8601DateFormatter()
@@ -30,7 +31,7 @@ extension JSONDecoder {
 }
 
 extension View {
-    /// Applies `.tabBarMinimizeBehavior(.onScrollDown)` if available on iOS.
+    /// Apply tab bar minimize behavior if available on iOS 26+
     @ViewBuilder
     func tabBarMinimizeOnScrollIfAvailable() -> some View {
         #if os(iOS)
@@ -46,6 +47,7 @@ extension View {
 }
 
 extension View {
+    /// Conditionally apply searchable modifier
     @ViewBuilder
     func searchableIf(_ condition: Bool, text: Binding<String>, prompt: String = "Search") -> some View {
         if condition {
@@ -56,24 +58,15 @@ extension View {
     }
 }
 
-/// Create a View extension that allows all Preview View objects to instantiate a fake data store to help with
-/// debugging.
 extension View {
-    /// Makes sure the Preview component comes with a faked datastore for populating UI elements.
-    ///
-    /// Don't want this running on real data so it's desirable to create a datastore like this. We don't always
-    /// need it though, so it's left as an optional extension.
+    /// Add fake datastore for Preview mode
     func withPreviewGrammarStore(mode: PreviewHelper = .normal) -> some View {
         PreviewHelper.withStore(mode: mode) { _ in
             self
         }
     }
 
-    /// Wraps the view in NavigationStack for components requiring navigation context.
-    ///
-    /// Many SwiftUI components rely on navigation context for proper rendering of
-    /// toolbars, titles, and navigation-dependent modifiers. This convenience method
-    /// ensures preview components have the necessary navigation environment.
+    /// Wrap view in NavigationStack for preview components
     func withPreviewNavigation() -> some View {
         NavigationStack {
             self
